Created the Chassis Class
Added the references to the navX file path so is useable, and then
created and implemented Chassis.java based on the Chassis.h and .cpp in
Geneva, duplicating the code into Java. Made all non-overriden methods
and variable static, and then added a getInstance to allow access to
systems that can't be static.