Removed OpenCV, Fixed navX-mxp, and Added Headlight
Removed the OpenCV reference, as vision code won't be ran on the roboRio
anymore. Fixed build.properties so as to make code referenceing the navX
build, by letting it find the file. Added a headlight subsystem, as well
as the relay port to RobotMap.java. Added two commands to handle turning
the headlight on and off, and bound them to buttons within OI.java to
allow the operator to turn it on and off.