Set up a "smart shifting" system.
Created a system that functions as a smart clutch while remaining fully
manual in the changing of gears. To achieve this a function was created
within Chassis that allows the switching of the SRXs to coast and brake
modes. The Shift Commands were edited to cut power to the motors, and
change to coast mode, then wait before shifting. After a slight delay it
then brings the power back based on the current position of the
Joysticks, and switches back into brake mode. Both of the delays are
configurable through preferences.